requirements.txt
langchain
python-dotenv
langchain-groq
langchain-community
streamlit
langchain-ollama
ollama
cohere